grpc/services/user-service/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.1,<=4.21.12
mysql-connector-python==8.1.0
bcrypt==4.0.1
PyJWT==2.8.0
kafka-python==2.0.2
python-dotenv==1.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0